Настраивает работу контроллеров
